@@ -6,8 +6,14 @@
     "type": "symfony-bundle",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0",
-        "symfony/framework-bundle": ">=2.1.0,<2.3-dev"
+        "php": ">=7.0",
+        "symfony/framework-bundle": "~3.0",
+        "symfony/security-bundle": "~3.0",
+        "zendframework/zend-ldap": "~2.0"
+    },
+     "require-dev": {
+        "phpunit/phpunit": "~6.0",
+        "friendsofphp/php-cs-fixer": "~2.0"
     },
     "autoload": {
         "psr-0": { "OpenSky\\Bundle\\LdapBundle": "" }
