@@ -2,13 +2,12 @@
 
 namespace OpenSky\Bundle\LdapBundle\DependencyInjection\Security;
 
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
-
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
+use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 
 /**
  * HttpBasicPreAuthenticatedFactory creates services for HTTP basic
@@ -30,19 +29,19 @@ class HttpBasicPreAuthenticatedFactory implements SecurityFactoryInterface
 
         $listener = new Definition(
             '%opensky_ldap.authentication.listener.basic_pre_auth.class%',
-            array(
+            [
                 new Reference('security.context'),
                 new Reference('security.authentication.manager'),
                 $id,
                 new Reference('logger', ContainerBuilder::IGNORE_ON_INVALID_REFERENCE),
-            )
+            ]
         );
-        $listener->addTag('monolog.logger', array('channel' => 'security'));
+        $listener->addTag('monolog.logger', ['channel' => 'security']);
 
         $listenerId = 'opensky_ldap.authentication.listener.basic_pre_auth.'.$id;
         $container->setDefinition($listenerId, $listener);
 
-        return array($provider, $listenerId, $defaultEntryPoint);
+        return [$provider, $listenerId, $defaultEntryPoint];
     }
 
     /**
