@@ -3,9 +3,13 @@
 namespace OpenSky\Bundle\LdapBundle\Tests\Security;
 
 use OpenSky\Bundle\LdapBundle\Security\HttpBasicPreAuthenticatedListener;
+use PHPUnit\Framework\TestCase;
+use ReflectionMethod;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
-class HttpBasicPreAuthenticatedListenerTest extends \PHPUnit_Framework_TestCase
+class HttpBasicPreAuthenticatedListenerTest extends TestCase
 {
     private $listener;
     private $method;
@@ -13,12 +17,12 @@ class HttpBasicPreAuthenticatedListenerTest extends \PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->listener = new HttpBasicPreAuthenticatedListener(
-            $this->getMock('Symfony\Component\Security\Core\SecurityContext', array(), array(), '', false, false),
-            $this->getMock('Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface'),
+            $this->createMock(TokenStorageInterface::class),
+            $this->createMock(AuthenticationManagerInterface::class),
             'ldap.provider'
         );
 
-        $this->method = new \ReflectionMethod($this->listener, 'getPreAuthenticatedData');
+        $this->method = new ReflectionMethod($this->listener, 'getPreAuthenticatedData');
         $this->method->setAccessible(true);
     }
 
@@ -27,27 +31,27 @@ class HttpBasicPreAuthenticatedListenerTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetPreAuthenticatedData($serverParams, $expectedData)
     {
-        $request = new Request(array(), array(), array(), array(), array(), $serverParams);
+        $request = new Request([], [], [], [], [], $serverParams);
 
         $this->assertEquals($expectedData, $this->method->invoke($this->listener, $request));
     }
 
     public function provideTestGetPreAuthenticatedData()
     {
-        return array(
-            array(
-                array('PHP_AUTH_USER' => 'username', 'PHP_AUTH_PW' => 'password'),
-                array('username', 'password'),
-            ),
-            array(
-                array('PHP_AUTH_USER' => 'username'),
-                array('username', ''),
-            ),
-        );
+        return [
+            [
+                ['PHP_AUTH_USER' => 'username', 'PHP_AUTH_PW' => 'password'],
+                ['username', 'password'],
+            ],
+            [
+                ['PHP_AUTH_USER' => 'username'],
+                ['username', ''],
+            ],
+        ];
     }
 
     /**
-     * @expectedException Symfony\Component\Security\Core\Exception\BadCredentialsException
+     * @expectedException \Symfony\Component\Security\Core\Exception\BadCredentialsException
      */
     public function testGetPreAuthenticatedDataBadCredentials()
     {
